Move app in own file in server, remove unsued func in ThankYouView
