+++ .gitignore
@@ -14,6 +14,7 @@ build/
 
 # misc
 keys/
+dump/
 .DS_Store
 .env
 .env.local
