+++ package.json
@@ -8,7 +8,7 @@
     "server": "nodemon ./server/index.js",
     "client": "npm start --prefix client",
     "dev": "concurrently \"npm run server\" \"npm run client\"",
-    "test": "jest ./server",
+    "test": "jest ./server --maxWorkers=1 --detectOpenHandles --forceExit",
     "heroku-postbuild": "NPM_CONFIG_PRODUCTION=false npm install --prefix client && npm run build --prefix client"
   },
   "repository": {
