+++ server/__tests__/itemModel.test.js
@@ -13,10 +13,6 @@ beforeAll(async () => {
 
 beforeEach(async () => {
   const collections = await mongoose.connection.db.collections();
-
-  for (let collection of collections) {
-    await collection.deleteMany({});
-  }
 });
 
 test("Item model schema should be present", () => {
@@ -24,6 +20,5 @@ test("Item model schema should be present", () => {
 });
   
 afterAll(async () => {
-  await mongoose.connection.dropDatabase()
   await mongoose.connection.close();
 });
