+++ server/__tests__/routes.test.js
@@ -1,7 +1,20 @@
 global.TextEncoder = require("util").TextEncoder;
 global.TextDecoder = require("util").TextDecoder;
-const { app } = require("../index");
+const { MongoMemoryServer } = require('mongodb-memory-server');
+const mongoose = require('mongoose');
+const { app } = require("../app");
 const request = require("supertest");
+
+beforeAll(async () => {
+  mongod = await MongoMemoryServer.create();
+  const mongoUri = mongod.getUri();
+
+  await mongoose.connect(mongoUri);
+});
+
+beforeEach(async () => {
+  const collections = await mongoose.connection.db.collections();
+});
 
 describe("Item get routes", () => {
   test("getItems should get all items and return status 200", done => {
@@ -17,4 +30,8 @@ describe("Item get routes", () => {
     .expect('Content-Type', /application\/json/)
     .expect(200, done);
   });
+});
+  
+afterAll(async () => {
+  await mongoose.connection.close();
 });
